requirements.txt
pyrogram
TgCrypto
python-dotenv
ffmpeg-python
